+++ src/Client/Client.h
@@ -100,6 +100,4 @@ public:
 
         return nullptr;
     }
-
-    void onPostSetup();
 };